Introduce RLTrainer class and move common code
